Add main.js and change 'Second Goal' in README
On new branch 'index_edit--script', <script> was added to the end
of 'index.html' to link to the newly created 'main.js'. This is
deemed necessary because JavaScript may be required to properly
scale the puzzle. An email has been sent to Jessica Kassel for
confirmation on whether or not this is allowed.

The 'Second Goal' in the README now also has a new edit stating the
time of this change.