Create scaling ability with SASS
After some research, I found a way to implement the scaling ability
using SASS instead of JavaScript. The functionality was implemented,
and a line was added to the 'Second Goal' of README recording the
change.

The only styling that is left to be done is to make the text within
each tile perfectly centered. I will first attempt to do this with
DOM manipulation.