Add logic for showing / swapping tiles
Coded function to show tiles and swap them if the blank tile is
directly above, below, or to either side.

The game is set upon page load (if a puzzle is available), so there
wasn't a need to add the blank tile outside of the function. That
code was therefore trimmed from 'main.js'. A note was added to 'Third
Goal' in README stating the change.

A unique class was also added to the blank tile so that its appearance
can mimic the requested form shown in the given images. (Done via
JavaScript)