Code different puzzle formats
The 3x3 and 4x4 puzzle capabilities were added. This was done via
two small objects for each format and replacing the hard-coded
dimensions with a replaceable parameter.

Next, event listeners will be added to the buttons in the option bar
to make it possible to switch between game modes. This change will
be reflected in the README.