Create functionality to select puzzle format
Added event listeners and tweaked game's logic to allow option bar
buttons to select a puzzle format. Specifically, an extra parameter
was added to 'playGame()' for affecting the dimensions of each puzzle.

The 'Fourth Goal' section in README was also updated to reflect the
new step in this process (adding event listeners).